Move MetadataAndProtocolEntries to out package
