+++ plugin/trino-delta-lake/src/main/java/io/trino/plugin/deltalake/transactionlog/MetadataAndProtocolEntries.java
@@ -11,10 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.deltalake.transactionlog.checkpoint;
-
-import io.trino.plugin.deltalake.transactionlog.MetadataEntry;
-import io.trino.plugin.deltalake.transactionlog.ProtocolEntry;
+package io.trino.plugin.deltalake.transactionlog;
 
 import java.util.Optional;
